fix(webhook): only send one response to client
during webhook processing some of the helper function can set a status
response for the client while the callee is also doing so. this can
cause a warning that the status is being set multiple times when they
differ. this change puts the responsibility on the caller of the helper
functions to return a response to the client - once.