fix(webhook): only send one response to client (#1154)
* fix(webhook): only send one response to client

during webhook processing some of the helper function can set a status
response for the client while the callee is also doing so. this can
cause a warning that the status is being set multiple times when they
differ. this change puts the responsibility on the caller of the helper
functions to return a response to the client - once.

* improve wording

---------

Co-authored-by: Easton Crupper <65553218+ecrupper@users.noreply.github.com>